Working example
Still has a segfault when the window is closed. Traced this to happening during the drop of `wgpu_hal::vulkan::DeviceShared`, specifically seems to happen when dropping the `instance` field, implying that the Arc is not in a valid state. Might be a bug in wgpu_hal. Interestingly, the order of the fields in AppState matters: the segfault only happens when `surface` is the last field to be dropped. Ignoring for now as it only happens when the window is closed, but needs investigating later.